+++ packages/dashboard-api/internal/handlers/utils_templates_list.go
@@ -106,24 +106,6 @@ func formatTemplatesCursor(sort templatesSort, value, id string) string {
 	return fmt.Sprintf("%s|%s|%s", sort, value, id)
 }
 
-func cursorTime(v *string) (*time.Time, error) {
-	if v == nil {
-		return nil, nil
-	}
-
-	t, err := time.Parse(time.RFC3339Nano, *v)
-	if err != nil {
-		return nil, fmt.Errorf("%w: %w", errInvalidTemplatesCursor, err)
-	}
-
-	return &t, nil
-}
-
-var (
-	maxCursorTime = time.Date(9999, 12, 31, 23, 59, 59, 0, time.UTC)
-	minCursorTime = time.Time{}
-)
-
 func timeCursor(ts *time.Time, id *string, desc bool) (time.Time, string) {
 	if ts != nil && id != nil {
 		return *ts, *id
